requirements.txt
opencv-python-headless
scikit-learn
scikit-image
tqdm
torch
torchvision
Pillow



